- Updated code changes for pcf exchange module
