+++ modules/pcf-exchange/pom.xml
@@ -15,6 +15,24 @@
 			<artifactId>sde-common</artifactId>
 			<version>0.0.1</version>
 		</dependency>
+		
+		<dependency>
+			<groupId>org.eclipse.tractusx</groupId>
+			<artifactId>portal</artifactId>
+			<version>0.0.1</version>
+		</dependency>
+
+		<dependency>
+			<groupId>org.eclipse.tractusx</groupId>
+			<artifactId>edc</artifactId>
+			<version>0.0.1</version>
+		</dependency>
+		
+		<dependency>
+			<groupId>org.eclipse.tractusx</groupId>
+			<artifactId>digital-twins</artifactId>
+			<version>0.0.1</version>
+		</dependency>
 	</dependencies>
 	<build>
 		<plugins>
