+++ modules/pcf-exchange/src/main/java/org/eclipse/tractusx/sde/pcfexchange/repository/PcfRequestRepository.java
@@ -1,8 +1,8 @@
-package org.eclipse.tractusx.sde.core.pcf.repository;
+package org.eclipse.tractusx.sde.pcfexchange.repository;
 
 import java.util.Optional;
 
-import org.eclipse.tractusx.sde.core.pcf.entity.PcfRequestEntity;
+import org.eclipse.tractusx.sde.pcfexchange.entity.PcfRequestEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PcfRequestRepository extends JpaRepository<PcfRequestEntity, String> {
