+++ modules/sde-core/pom.xml
@@ -178,6 +178,11 @@
 		</dependency>
 		<dependency>
 			<groupId>org.eclipse.tractusx</groupId>
+			<artifactId>pcf-exchange</artifactId>
+			<version>0.0.1</version>
+		</dependency>
+		<dependency>
+			<groupId>org.eclipse.tractusx</groupId>
 			<artifactId>sde-agent</artifactId>
 			<version>0.0.1</version>
 		</dependency>
